+++ internal/aghhttp/json.go
@@ -1,4 +1,4 @@
-package websvc
+package aghhttp
 
 import (
 	"encoding/json"
@@ -7,7 +7,6 @@ import (
 	"strconv"
 	"time"
 
-	"github.com/AdguardTeam/AdGuardHome/internal/aghhttp"
 	"github.com/AdguardTeam/golibs/httphdr"
 	"github.com/AdguardTeam/golibs/log"
 )
@@ -87,28 +86,27 @@ func (t *JSONTime) UnmarshalJSON(b []byte) (err error) {
 	return nil
 }
 
-// writeJSONOKResponse writes headers with the code 200 OK, encodes v into w,
-// and logs any errors it encounters.  r is used to get additional information
-// from the request.
-func writeJSONOKResponse(w http.ResponseWriter, r *http.Request, v any) {
-	writeJSONResponse(w, r, v, http.StatusOK)
-}
-
-// writeJSONResponse writes headers with code, encodes v into w, and logs any
-// errors it encounters.  r is used to get additional information from the
+// WriteJSONResponse writes headers with the code, encodes resp into w, and logs
+// any errors it encounters.  r is used to get additional information from the
 // request.
-func writeJSONResponse(w http.ResponseWriter, r *http.Request, v any, code int) {
-	// TODO(a.garipov): Put some of these to a middleware.
+func WriteJSONResponse(w http.ResponseWriter, r *http.Request, code int, resp any) {
 	h := w.Header()
-	h.Set(httphdr.ContentType, aghhttp.HdrValApplicationJSON)
-	h.Set(httphdr.Server, aghhttp.UserAgent())
+	h.Set(httphdr.ContentType, HdrValApplicationJSON)
+	h.Set(httphdr.Server, UserAgent())
 
 	w.WriteHeader(code)
 
-	err := json.NewEncoder(w).Encode(v)
+	err := json.NewEncoder(w).Encode(resp)
 	if err != nil {
-		log.Error("websvc: writing resp to %s %s: %s", r.Method, r.URL.Path, err)
+		log.Error("aghhttp: writing json resp to %s %s: %s", r.Method, r.URL.Path, err)
 	}
+}
+
+// WriteJSONResponseOK writes headers with the code 200 OK, encodes v into w,
+// and logs any errors it encounters.  r is used to get additional information
+// from the request.
+func WriteJSONResponseOK(w http.ResponseWriter, r *http.Request, v any) {
+	WriteJSONResponse(w, r, http.StatusOK, v)
 }
 
 // ErrorCode is the error code as used by the HTTP API.  See the ErrorCode
@@ -131,14 +129,14 @@ type HTTPAPIErrorResp struct {
 	Msg  string    `json:"msg"`
 }
 
-// writeJSONErrorResponse encodes err as a JSON error into w, and logs any
+// WriteJSONResponseError encodes err as a JSON error into w, and logs any
 // errors it encounters.  r is used to get additional information from the
 // request.
-func writeJSONErrorResponse(w http.ResponseWriter, r *http.Request, err error) {
-	log.Error("websvc: %s %s: %s", r.Method, r.URL.Path, err)
+func WriteJSONResponseError(w http.ResponseWriter, r *http.Request, err error) {
+	log.Error("aghhttp: writing json error to %s %s: %s", r.Method, r.URL.Path, err)
 
-	writeJSONResponse(w, r, &HTTPAPIErrorResp{
+	WriteJSONResponse(w, r, http.StatusUnprocessableEntity, &HTTPAPIErrorResp{
 		Code: ErrorCodeTMP000,
 		Msg:  err.Error(),
-	}, http.StatusUnprocessableEntity)
+	})
 }