+++ internal/next/websvc/dns.go
@@ -7,6 +7,7 @@ import (
 	"net/netip"
 	"time"
 
+	"github.com/AdguardTeam/AdGuardHome/internal/aghhttp"
 	"github.com/AdguardTeam/AdGuardHome/internal/next/dnssvc"
 )
 
@@ -17,13 +18,13 @@ import (
 type ReqPatchSettingsDNS struct {
 	// TODO(a.garipov): Add more as we go.
 
-	Addresses           []netip.AddrPort `json:"addresses"`
-	BootstrapServers    []string         `json:"bootstrap_servers"`
-	UpstreamServers     []string         `json:"upstream_servers"`
-	DNS64Prefixes       []netip.Prefix   `json:"dns64_prefixes"`
-	UpstreamTimeout     JSONDuration     `json:"upstream_timeout"`
-	BootstrapPreferIPv6 bool             `json:"bootstrap_prefer_ipv6"`
-	UseDNS64            bool             `json:"use_dns64"`
+	Addresses           []netip.AddrPort     `json:"addresses"`
+	BootstrapServers    []string             `json:"bootstrap_servers"`
+	UpstreamServers     []string             `json:"upstream_servers"`
+	DNS64Prefixes       []netip.Prefix       `json:"dns64_prefixes"`
+	UpstreamTimeout     aghhttp.JSONDuration `json:"upstream_timeout"`
+	BootstrapPreferIPv6 bool                 `json:"bootstrap_prefer_ipv6"`
+	UseDNS64            bool                 `json:"use_dns64"`
 }
 
 // HTTPAPIDNSSettings are the DNS settings as used by the HTTP API.  See the
@@ -31,13 +32,13 @@ type ReqPatchSettingsDNS struct {
 type HTTPAPIDNSSettings struct {
 	// TODO(a.garipov): Add more as we go.
 
-	Addresses           []netip.AddrPort `json:"addresses"`
-	BootstrapServers    []string         `json:"bootstrap_servers"`
-	UpstreamServers     []string         `json:"upstream_servers"`
-	DNS64Prefixes       []netip.Prefix   `json:"dns64_prefixes"`
-	UpstreamTimeout     JSONDuration     `json:"upstream_timeout"`
-	BootstrapPreferIPv6 bool             `json:"bootstrap_prefer_ipv6"`
-	UseDNS64            bool             `json:"use_dns64"`
+	Addresses           []netip.AddrPort     `json:"addresses"`
+	BootstrapServers    []string             `json:"bootstrap_servers"`
+	UpstreamServers     []string             `json:"upstream_servers"`
+	DNS64Prefixes       []netip.Prefix       `json:"dns64_prefixes"`
+	UpstreamTimeout     aghhttp.JSONDuration `json:"upstream_timeout"`
+	BootstrapPreferIPv6 bool                 `json:"bootstrap_prefer_ipv6"`
+	UseDNS64            bool                 `json:"use_dns64"`
 }
 
 // handlePatchSettingsDNS is the handler for the PATCH /api/v1/settings/dns HTTP
@@ -53,7 +54,7 @@ func (svc *Service) handlePatchSettingsDNS(w http.ResponseWriter, r *http.Reques
 
 	err := json.NewDecoder(r.Body).Decode(&req)
 	if err != nil {
-		writeJSONErrorResponse(w, r, fmt.Errorf("decoding: %w", err))
+		aghhttp.WriteJSONResponseError(w, r, fmt.Errorf("decoding: %w", err))
 
 		return
 	}
@@ -71,7 +72,7 @@ func (svc *Service) handlePatchSettingsDNS(w http.ResponseWriter, r *http.Reques
 	ctx := r.Context()
 	err = svc.confMgr.UpdateDNS(ctx, newConf)
 	if err != nil {
-		writeJSONErrorResponse(w, r, fmt.Errorf("updating: %w", err))
+		aghhttp.WriteJSONResponseError(w, r, fmt.Errorf("updating: %w", err))
 
 		return
 	}
@@ -79,17 +80,17 @@ func (svc *Service) handlePatchSettingsDNS(w http.ResponseWriter, r *http.Reques
 	newSvc := svc.confMgr.DNS()
 	err = newSvc.Start()
 	if err != nil {
-		writeJSONErrorResponse(w, r, fmt.Errorf("starting new service: %w", err))
+		aghhttp.WriteJSONResponseError(w, r, fmt.Errorf("starting new service: %w", err))
 
 		return
 	}
 
-	writeJSONOKResponse(w, r, &HTTPAPIDNSSettings{
+	aghhttp.WriteJSONResponseOK(w, r, &HTTPAPIDNSSettings{
 		Addresses:           newConf.Addresses,
 		BootstrapServers:    newConf.BootstrapServers,
 		UpstreamServers:     newConf.UpstreamServers,
 		DNS64Prefixes:       newConf.DNS64Prefixes,
-		UpstreamTimeout:     JSONDuration(newConf.UpstreamTimeout),
+		UpstreamTimeout:     aghhttp.JSONDuration(newConf.UpstreamTimeout),
 		BootstrapPreferIPv6: newConf.BootstrapPreferIPv6,
 		UseDNS64:            newConf.UseDNS64,
 	})